# services/academic-service/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Academic Service API
  version: 1.0.0
  description: |
    The Academic Service is one of the microservices of the school ERP system. It owns courses,
    programs, enrollment, class sessions and attendance, grades and GPA, the at-risk-student flag,
    exam scheduling, grade appeals and transcript export, and its own database (`academic_db`).
    Other services (Auth, Finance, HR, Notifications) never read its tables; they get academic data
    through this REST API (normally via the API gateway) or through the RabbitMQ events it publishes
    (`academic.student.enrolled`, `academic.grade.published`, `academic.student.at_risk_flagged`).

    This spec describes the service **as implemented**. The written contract is
    `docs/api-contracts/academic-service.md`; the two must be updated together.

    **Conventions**
    - Every endpoint is under `/api/v1/academic`, the prefix the API gateway forwards to this service.
      Only `GET /health` is unprefixed (for Docker health checks).
    - No authentication yet; it will be added once the Auth service and gateway headers are in place.
    - Errors are JSON `{ "error": "message" }`. Some add fields: `code` (appeals),
      `missingPrerequisites` (enrollment 409) and `conflicts` (exam 409).
    - A request body that isn't valid JSON returns `400 { "error": "Request body is not valid JSON" }`.
    - Ids in paths must be positive integers; anything else returns `400`.
...
servers:
...
paths:
...
  /api/v1/academic/programs:
    post:
      tags: [Programs]
      summary: Create a program
      operationId: createProgram
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProgramInput' }
      responses:
        '201':
          description: Program created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Program' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: A program with this name already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'A program named "BSc Computer Science" already exists' }
        '500': { $ref: '#/components/responses/ServerError' }
    get:
      tags: [Programs]
      summary: List all programs
      description: Ordered by id. Returns `[]` if there are none.
      operationId: listPrograms
      responses:
        '200':
          description: All programs.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Program' }
        '500': { $ref: '#/components/responses/ServerError' }

  # ---------------------------------------------------------------- Courses
  /api/v1/academic/courses:
    post:
      tags: [Courses]
      summary: Create a course (optionally with prerequisites)
      description: |
        The course and its prerequisite links are saved in one transaction: if linking fails, the
        course isn't saved either. Duplicate ids in `prerequisite_ids` are ignored.
      operationId: createCourse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseInput' }
      responses:
        '201':
          description: Course created, with its prerequisites resolved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseWithPrerequisites' }
        '400':
          description: Missing or invalid field, `program_id` doesn't exist, or unknown prerequisite id(s).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                missing: { value: { error: 'Missing required field(s): code, credit_hours' } }
                unknownPrerequisite: { value: { error: 'prerequisite course id(s) do not exist: 77' } }
        '409':
          description: A course with this `code` already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'A course with code "CS101" already exists' }
        '500': { $ref: '#/components/responses/ServerError' }
    get:
      tags: [Courses]
      summary: List all courses
      description: Ordered by id. Prerequisites are not included; use GET `/api/v1/academic/courses/{id}`.
      operationId: listCourses
      responses:
        '200':
          description: All courses.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Course' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/academic/enrollments:
    post:
      tags: [Enrollment]
      summary: Enroll a student in a course for a semester
      description: |
        Checks that every prerequisite was taken (not dropped) in a semester that **ended before
        this semester starts**. Grades aren't consulted yet: "enrolled" counts as "completed".
        After saving, publishes `academic.student.enrolled` (with `tuitionAmount` from the
        `TUITION_AMOUNT` setting). If RabbitMQ is down, the enrollment still succeeds and the
        missed event is logged.
      operationId: createEnrollment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollmentInput' }
      responses:
        '201':
          description: Enrolled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Enrollment' }
        '400':
          description: Missing or invalid field, or the student, course or semester doesn't exist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Prerequisites not met (with `missingPrerequisites`), or already enrolled in this course this semester.
          content:
            application/json:
              schema:
                # anyOf, not oneOf: a prerequisite-conflict body also satisfies the plain Error schema.
                anyOf:
                  - $ref: '#/components/schemas/PrerequisiteConflict'
                  - $ref: '#/components/schemas/Error'
              examples:
                prerequisitesNotMet:
                  value:
                    error: 'Prerequisites not met for CS201. Missing: CS102 (Discrete Math)'
                    missingPrerequisites: [{ id: 2, code: CS102, title: Discrete Math }]
                alreadyEnrolled:
                  value: { error: 'Student 1 is already enrolled in course 1 for semester 1' }
        '500':
          description: Server error, or `TUITION_AMOUNT` isn't configured (nothing is saved in that case).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/v1/academic/sessions:
    post:
      tags: [Attendance]
      summary: Create a class session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionInput' }
      responses:
        '201':
          description: Session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClassSession' }
        '400':
          description: Missing or invalid field, `endTime` not after `startTime`, course/semester doesn't exist, or date outside the semester.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'sessionDate 2027-01-10 is outside Fall 2026 (2026-09-01 to 2026-12-20)' }
        '409':
          description: A session for this course, semester, date and start time already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500': { $ref: '#/components/responses/ServerError' }
  /api/v1/academic/attendance:
    post:
      tags: [Attendance]
      summary: Record attendance for a whole session
      description: |
        All-or-nothing: if any record fails, nothing is saved. Every student must be actively
        enrolled in the session's course and semester. After saving, each listed student's at-risk
        status is re-checked (which may publish `academic.student.at_risk_flagged`).
      operationId: recordAttendance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceInput' }
      responses:
        '201':
          description: Saved. Returns the session and its full register.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionAttendance' }
        '400':
          description: Invalid record (the error names its index), a student listed twice, session doesn't exist, or student(s) not enrolled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                badRecord: { value: { error: 'records[1].status must be one of: present, absent, late' } }
                notEnrolled: { value: { error: "studentId(s) not enrolled in this session's course and semester: 3, 99" } }
        '409':
          description: Some students are already marked for this session (named in the error). Nothing was saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'Attendance already recorded for studentId(s) 1 in session 1; nothing was saved' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/academic/grades:
    post:
      tags: [Grades]
      summary: Record or correct a draft grade
      description: |
        One grade per student, course and semester. The first POST creates a draft (`201`); posting
        again corrects the draft (`200`). A **published** grade can't be changed this way (`409`);
        the only way to change it is an approved grade appeal.
      operationId: recordGrade
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GradeInput' }
      responses:
        '200':
          description: Existing draft corrected.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grade' }
        '201':
          description: New draft created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Grade' }
        '400':
          description: Missing or invalid field, student/course/semester doesn't exist, or the student isn't enrolled in that course that semester.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: The grade is already published and can't be changed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: "Grade 1 for this student, course and semester is already published and can't be changed" }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/v1/academic/exams:
    post:
      tags: [Exams]
      summary: Schedule an exam
      description: |
        Two exams conflict if they're in the same room (case-insensitive, surrounding spaces ignored),
        on the same date, and their half-open time ranges overlap: `[s1, e1)` overlaps `[s2, e2)` when
        `s1 < e2 AND s2 < e1`. Back-to-back exams do **not** conflict. Bookings for the same room are
        serialised with a lock, so simultaneous requests can't double-book it.
      operationId: createExam
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExamInput' }
      responses:
        '201':
          description: Exam scheduled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Exam' }
        '400':
          description: Missing or invalid field, `endTime` not after `startTime`, course/semester doesn't exist, or date outside the semester.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Overlaps existing exam(s) in that room on that date. Nothing was saved.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExamConflict' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/RoomLockBusy' }
    get:
      tags: [Exams]
      summary: List exams
      description: Ordered by date, start time, room. Both filters are optional and can be combined.
      operationId: listExams
      parameters:
        - name: semesterId
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - name: courseId
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Matching exams.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Exam' }
        '400':
          description: A filter is present but isn't a positive integer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: 'courseId must be a positive integer' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
...
  responses:
    BadRequest:
      description: Invalid input (e.g. a missing field, or an id that isn't a positive integer), or a body that isn't valid JSON.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'id must be a positive integer' }
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RoomLockBusy:
      description: The room's booking lock was busy for more than 5 seconds. Temporary; retry.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'Timed out waiting for the booking lock on room "Hall A"; please retry' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable message. }
...
    PrerequisiteConflict:
      type: object
      required: [error, missingPrerequisites]
      properties:
        error: { type: string }
        missingPrerequisites:
          type: array
          items:
            type: object
            required: [id, code, title]
            properties:
              id: { type: integer }
              code: { type: string }
              title: { type: string }
    ExamConflict:
      type: object
      required: [error, conflicts]
      properties:
        error:
          type: string
          example: 'Room "Hall A" is already booked on 2026-12-10 at an overlapping time: CS101 09:00-11:00 (exam 1). The exam was not saved.'
        conflicts:
          type: array
          description: Every existing exam the requested slot overlaps, ordered by start time.
          items: { $ref: '#/components/schemas/Exam' }

    # ---------- shared formats
    Date:
      type: string
      format: date
      example: '2026-12-10'
    TimeInput:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d(:[0-5]\d)?$'
      description: 24-hour time, `HH:MM` or `HH:MM:SS`.
      example: '09:00'
    Time:
      type: string
      pattern: '^\d{2}:\d{2}:\d{2}$'
      example: '09:00:00'
    GradeLetter:
      type: string
      enum: [A, B, C, D, F]
    GradePoints:
      type: number
      minimum: 0
      maximum: 4
      description: 0 to 4 with at most 2 decimal places. Must be a JSON number, not a string.
      example: 3.7

    # ---------- programs & courses
    ProgramInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, description: Leading/trailing spaces are trimmed. }
        description: { type: string, nullable: true }
    Program:
      type: object
      required: [id, name, description, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    CourseInput:
      type: object
      required: [program_id, code, title, credit_hours]
      properties:
        program_id: { type: integer, minimum: 1 }
        code: { type: string, minLength: 1, maxLength: 20 }
        title: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, nullable: true }
        credit_hours: { type: integer, minimum: 1, maximum: 255 }
        prerequisite_ids:
          type: array
          items: { type: integer, minimum: 1 }
          description: Optional. On PUT, omit to keep the current prerequisites; `[]` clears them.
    Course:
      type: object
      required: [id, program_id, code, title, description, credit_hours, created_at]
      properties:
        id: { type: integer }
        program_id: { type: integer }
        code: { type: string, example: CS201 }
        title: { type: string, example: Data Structures }
        description: { type: string, nullable: true }
        credit_hours: { type: integer, example: 4 }
        created_at: { type: string, format: date-time }
    CoursePrerequisite:
      type: object
      required: [id, code, title, credit_hours]
      properties:
        id: { type: integer }
        code: { type: string }
        title: { type: string }
        credit_hours: { type: integer }
    CourseWithPrerequisites:
      allOf:
        - $ref: '#/components/schemas/Course'
        - type: object
          required: [prerequisites]
          properties:
            prerequisites:
              type: array
              items: { $ref: '#/components/schemas/CoursePrerequisite' }

    # ---------- enrollment
    EnrollmentInput:
      type: object
      required: [studentId, courseId, semesterId]
      properties:
        studentId: { type: integer, minimum: 1 }
        courseId: { type: integer, minimum: 1 }
        semesterId: { type: integer, minimum: 1 }
    Enrollment:
      type: object
      required: [id, student_id, course_id, semester_id, status, enrolled_at]
      properties:
        id: { type: integer }
        student_id: { type: integer }
        course_id: { type: integer }
        semester_id: { type: integer }
        status: { type: string, enum: [enrolled, dropped] }
        enrolled_at: { type: string, format: date-time }
...
    SessionInput:
      type: object
      required: [courseId, semesterId, sessionDate, startTime, endTime]
      properties:
        courseId: { type: integer, minimum: 1 }
        semesterId: { type: integer, minimum: 1 }
        sessionDate: { $ref: '#/components/schemas/Date' }
        startTime: { $ref: '#/components/schemas/TimeInput' }
        endTime: { $ref: '#/components/schemas/TimeInput' }
    ClassSession:
      type: object
      required: [id, course_id, semester_id, session_date, start_time, end_time, created_at]
      properties:
        id: { type: integer }
        course_id: { type: integer }
        semester_id: { type: integer }
        session_date: { $ref: '#/components/schemas/Date' }
        start_time: { $ref: '#/components/schemas/Time' }
        end_time: { $ref: '#/components/schemas/Time' }
        created_at: { type: string, format: date-time }
    AttendanceStatus:
      type: string
      enum: [present, absent, late]
    AttendanceInput:
      type: object
      required: [sessionId, records]
      properties:
        sessionId: { type: integer, minimum: 1 }
        records:
          type: array
          minItems: 1
          description: One entry per student; the same student may not appear twice.
          items:
            type: object
            required: [studentId, status]
            properties:
              studentId: { type: integer, minimum: 1 }
              status: { $ref: '#/components/schemas/AttendanceStatus' }
    AttendanceRecord:
      type: object
      required: [id, student_id, first_name, last_name, status, recorded_at]
      properties:
        id: { type: integer }
        student_id: { type: integer }
        first_name: { type: string }
        last_name: { type: string }
        status: { $ref: '#/components/schemas/AttendanceStatus' }
        recorded_at: { type: string, format: date-time }
    SessionAttendance:
      type: object
      required: [session, records]
      properties:
        session: { $ref: '#/components/schemas/ClassSession' }
        records:
          type: array
          items: { $ref: '#/components/schemas/AttendanceRecord' }
...
    GradeInput:
      type: object
      required: [studentId, courseId, semesterId, gradeLetter, gradePoints]
      properties:
        studentId: { type: integer, minimum: 1 }
        courseId: { type: integer, minimum: 1 }
        semesterId: { type: integer, minimum: 1 }
        gradeLetter:
          type: string
          description: A-F; lowercase is accepted and stored in uppercase.
          example: B
        gradePoints: { $ref: '#/components/schemas/GradePoints' }
    Grade:
      type: object
      required: [id, student_id, course_id, semester_id, grade_letter, grade_points, published, published_at, created_at]
      properties:
        id: { type: integer }
        student_id: { type: integer }
        course_id: { type: integer }
        semester_id: { type: integer }
        grade_letter: { $ref: '#/components/schemas/GradeLetter' }
        grade_points: { type: number, example: 3 }
        published: { type: boolean }
        published_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
    ExamInput:
      type: object
      required: [courseId, semesterId, examDate, startTime, endTime, room]
      properties:
        courseId: { type: integer, minimum: 1 }
        semesterId: { type: integer, minimum: 1 }
        examDate: { $ref: '#/components/schemas/Date' }
        startTime: { $ref: '#/components/schemas/TimeInput' }
        endTime: { $ref: '#/components/schemas/TimeInput' }
        room:
          type: string
          minLength: 1
          maxLength: 50
          description: Trimmed; compared case-insensitively when checking for conflicts.
          example: Hall A
    Exam:
      type: object
      required: [id, course_id, course_code, course_title, semester_id, semester_name, exam_date, start_time, end_time, room, created_at]
      properties:
        id: { type: integer }
        course_id: { type: integer }
        course_code: { type: string, example: CS101 }
        course_title: { type: string }
        semester_id: { type: integer }
        semester_name: { type: string, example: Fall 2026 }
        exam_date: { $ref: '#/components/schemas/Date' }
        start_time: { $ref: '#/components/schemas/Time' }
        end_time: { $ref: '#/components/schemas/Time' }
        room: { type: string, example: Hall A }
        created_at: { type: string, format: date-time }

    # ---------- appeals
...